Update flask and break apart different sections into different files
